# Conflicts:
#	src/othello/interface.cpp
#	src/shogi/usi.cpp

+++ src/learn.hpp
@@ -1,24 +1,32 @@
 ﻿#ifndef MIACIS_LEARN_HPP
 #define MIACIS_LEARN_HPP
 
-#include"neural_network.hpp"
-#include<string>
-#include<chrono>
+#include "neural_network.hpp"
+#include <chrono>
+#include <string>
 
 //経過時間を取得する関数
 std::string elapsedTime(const std::chrono::steady_clock::time_point& start);
-double elapsedHours(const std::chrono::steady_clock::time_point& start);
+float elapsedHours(const std::chrono::steady_clock::time_point& start);
 
 //標準出力とファイルストリームに同時に出力するためのクラス
 //参考)https://aki-yam.hatenablog.com/entry/20080630/1214801872
 class dout {
 private:
     std::ostream &os1, &os2;
+
 public:
-    explicit dout (std::ostream &_os1, std::ostream &_os2) : os1(_os1), os2(_os2) {};
-    template <typename T>
-    dout& operator<< (const T &rhs)  { os1 << rhs;  os2 << rhs; return *this; };
-    dout& operator<< (std::ostream& (*__pf)(std::ostream&))  { __pf(os1); __pf(os2); return *this; };
+    explicit dout(std::ostream& _os1, std::ostream& _os2) : os1(_os1), os2(_os2){};
+    template<typename T> dout& operator<<(const T& rhs) {
+        os1 << rhs;
+        os2 << rhs;
+        return *this;
+    };
+    dout& operator<<(std::ostream& (*__pf)(std::ostream&)) {
+        __pf(os1);
+        __pf(os2);
+        return *this;
+    };
     /*!<  Interface for manipulators, such as \c std::endl and \c std::setw
       For more information, see ostream header */
 };
@@ -39,7 +47,8 @@ public:
 std::vector<LearningData> loadData(const std::string& file_path, bool data_augmentation);
 
 //validationを行う関数
-std::array<float, LOSS_TYPE_NUM> validation(NeuralNetwork nn, const std::vector<LearningData>& validation_data, uint64_t batch_size);
+std::array<float, LOSS_TYPE_NUM> validation(NeuralNetwork nn, const std::vector<LearningData>& validation_data,
+                                            uint64_t batch_size);
 
 //パラメータを初期化
 void initParams();
